Refactor config modules into more modular modules.
